Fix issue with actor checkpointing
If an actor is checkpointed, the instance that it's superseding can
still be running and may finish even though you reverted to the
checkpointed state. The problem is that when it finishes, all the new
actors that it spawned and new messages that it queued up to send out
will still get processed. This fix stops that from happening.

Also, this commit also clarifies some documentation.